requirements.txt
# PUO Memo Clean Architecture - Dependencies

# Core dependencies
mcp>=1.0.0
asyncpg>=0.29.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Optional AI features
google-generativeai>=0.3.2

# API Server dependencies
aiohttp>=3.9.0
aiohttp-cors>=0.7.0
